requirements.txt
requests==2.9.1
dj-database-url==0.4.0
Django==2.2.16
gunicorn==19.4.5
psycopg2-binary
whitenoise==2.0.6
django-ckeditor
pyimgur
